refactor(866): dissolve Humans.Web/Infrastructure/ into Web's own folders
The folder was a relic of deleted Humans.Infrastructure. Contents move to
Web-shaped homes with matching Humans.Web.* namespaces:

- Data/: SystemDbContext + factory (Migrations/System namespace line updated
  to match — the one sanctioned migration edit)
- Repositories/Admin/: admin diagnostics pair
- Services/: the 8 operational trackers + GateLoginThrottle +
  RateLimitRejectionAggregator
- Hosting/: section feature providers, dev-login exclusion, gate terminal seeder
- ModelBinders/: LocalDateTimeModelBinder
- Extensions/: AddHumansPersistence wiring, class renamed
  InfrastructureServiceCollectionExtensions -> PersistenceServiceCollectionExtensions
  (the stale namespace existed only to dodge the roll-call name collision)

Humans.Web no longer declares any Humans.Infrastructure.* namespace; the 20
files in Humans.Interfaces keep theirs until the 4b-iv Base rename. Docs and
freshness globs updated; design-rules §15's one stale path is left for the
§15 rewrite.

Tests: 5,362 passed / 0 failed / 5 skipped (baseline).
