+++ docs/architecture/data-model.md
@@ -6,7 +6,7 @@ This file is the **index and cross-cutting rule sheet** for the data model. Per-
 
 ## Entity index
 
-<!-- freshness:auto id="entity-index" prompt="Walk every entity class under src/Sections/*/Domain/ (configured under that section's own Data/ tree — search Data/ recursively, since some sections keep configuration classes directly under Data/ rather than Data/Configurations/, e.g. Consent and Email), plus the one non-section context, SystemDbContext in src/Humans.Web/Infrastructure/Data/. Humans.Domain and Humans.Infrastructure were deleted across G5, so there is no longer a second entity home to walk. For each, identify the owning section by finding the section doc whose Data Model section names the entity — that doc is either docs/sections/<X>.md or, for a section that has moved into its own project, src/Sections/Humans.<X>/Docs/<X>.md; docs/sections/_Index.md maps which is which. Do not assume a Holded*-prefixed entity belongs to Finance: Finance and Holded are separate sections owning separate tables in separate DbContexts. Build the entity index table with columns: Entity | Owning section | Notes. Preserve any per-row Notes column content the existing table already has — only update entity names and section links if they changed." -->
+<!-- freshness:auto id="entity-index" prompt="Walk every entity class under src/Sections/*/Domain/ (configured under that section's own Data/ tree — search Data/ recursively, since some sections keep configuration classes directly under Data/ rather than Data/Configurations/, e.g. Consent and Email), plus the one non-section context, SystemDbContext in src/Humans.Web/Data/. Humans.Domain and Humans.Infrastructure were deleted across G5, so there is no longer a second entity home to walk. For each, identify the owning section by finding the section doc whose Data Model section names the entity — that doc is either docs/sections/<X>.md or, for a section that has moved into its own project, src/Sections/Humans.<X>/Docs/<X>.md; docs/sections/_Index.md maps which is which. Do not assume a Holded*-prefixed entity belongs to Finance: Finance and Holded are separate sections owning separate tables in separate DbContexts. Build the entity index table with columns: Entity | Owning section | Notes. Preserve any per-row Notes column content the existing table already has — only update entity names and section links if they changed." -->
 
 | Entity | Owning section | Notes |
 |--------|---------------|-------|
