+++ docs/features/scanner/gate-terminal-login.md
@@ -3,8 +3,8 @@
   src/Sections/Humans.Tickets/Controllers/TicketsOnsiteAdminController.cs
   src/Humans.Web/Views/Account/GateLogin.cshtml
   src/Humans.Web/Views/Tickets/Admin/Gate.cshtml
-  src/Humans.Web/Infrastructure/GateTerminalAccountSeeder.cs
-  src/Humans.Web/Infrastructure/GateLoginThrottle.cs
+  src/Humans.Web/Hosting/GateTerminalAccountSeeder.cs
+  src/Humans.Web/Services/GateLoginThrottle.cs
   src/Humans.Domain/Constants/SystemUserIds.cs
   src/Humans.UI/Authorization/PolicyNames.cs
   src/Humans.Web/Authorization/AuthorizationPolicyExtensions.cs
