+++ docs/sections/G5-SECTION-TEMPLATE.md
@@ -1155,7 +1155,7 @@ Git Bash.)
      two things. First, MVC's `ViewComponentConventions.IsComponent` requires `IsPublic`, so an
      `internal` component is silently never discovered — exactly the hazard
      `SectionControllerFeatureProvider` exists for on the controller side. The counterpart is
-     `Humans.Web/Infrastructure/SectionViewComponentFeatureProvider`: a second
+     `Humans.Web/Hosting/SectionViewComponentFeatureProvider`: a second
      `IApplicationFeatureProvider<ViewComponentFeature>` pass (the base one is not virtual and
      `ViewComponentConventions` is internal to MVC) that adds non-public components from
      discovered section assemblies. Write it once; every later section with a
@@ -1731,7 +1731,7 @@ Git Bash.)
       controller needs no edit at all. One file, two rows about the same controller, only one
       of which is a `typeof` (proven: Debug, `DebugController.DbVersion`).
       **And a third shape, which is not in `tests/` at all: Shell's own production code.**
-      `Humans.Web/Infrastructure/DevLoginControllerExclusionProvider` removes
+      `Humans.Web/Hosting/DevLoginControllerExclusionProvider` removes
       `typeof(Controllers.DevLoginController)` from MVC's controller feature in Production —
       the thing that keeps the dev sign-in page out of prod. It cannot move into the section
       (`Program.cs` constructs it by name, which would make it a public section type) and it
