+++ src/Humans.Web/Controllers/AccountController.cs
@@ -8,7 +8,7 @@ using NodaTime;
 using Humans.Application.Extensions;
 using Humans.Domain.Constants;
 using Humans.UI;
-using Humans.Web.Infrastructure;
+using Humans.Web.Services;
 using Humans.Users.Contracts;
 
 namespace Humans.Web.Controllers;
