+++ src/Humans.Web/Controllers/TicketsGateAdminController.cs
@@ -1,6 +1,6 @@
 using Humans.UI.Authorization;
 using Humans.UI.Controllers;
-using Humans.Web.Infrastructure;
+using Humans.Web.Hosting;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Humans.Users.Contracts;
