+++ src/Humans.Web/Extensions/InfrastructureServiceCollectionExtensions.cs
@@ -14,6 +14,7 @@ using Humans.Tickets.Jobs;
 using Humans.Infrastructure.Caching;
 using Humans.Infrastructure.Configuration;
 using Humans.Infrastructure.Services;
+using Humans.Web.Services;
 using Humans.Issues.Jobs;
 using Humans.Notifications.Jobs;
 using Humans.Web.Extensions.Infrastructure;
