+++ src/Humans.Web/Program.cs
@@ -23,7 +23,7 @@ using Humans.Web.Extensions;
 using Microsoft.Extensions.Caching.Memory;
 using Humans.Infrastructure.Data;
 using Humans.Infrastructure.Hosting;
-using Humans.Infrastructure.Services;
+using Humans.Web.Services;
 using Humans.Web.Authorization;
 using Humans.Web.Health;
 using Humans.CityPlanning.Contracts;
@@ -34,7 +34,9 @@ using Humans.Infrastructure.Logging;
 using Humans.UI.Extensions;
 using Serilog;
 using Serilog.Events;
-using Humans.Web.Infrastructure;
+using Humans.Web.Hosting;
+using Humans.Web.ModelBinders;
+using Humans.Web.Data;
 using Humans.Users.Contracts;
 
 var builder = WebApplication.CreateBuilder(args);
