+++ tests/Humans.Integration.Tests/Infrastructure/PhysicalDefaultParityTests.cs
@@ -1,5 +1,6 @@
 using AwesomeAssertions;
 using Humans.Infrastructure.Data;
+using Humans.Web.Data;
 using Humans.Infrastructure.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata;
