+++ tests/Humans.Users.Tests/Infrastructure/TestMetrics.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging.Abstractions;
 using NSubstitute;
 using Humans.Application.Interfaces;
-using Humans.Infrastructure.Services;
+using Humans.Web.Services;
 
 namespace Humans.Users.Tests.Infrastructure;
