+++ tests/Humans.Web.Tests/Services/ClientStatsTrackerTests.cs
@@ -1,6 +1,6 @@
 using AwesomeAssertions;
 using Humans.Application.Interfaces;
-using Humans.Infrastructure.Services;
+using Humans.Web.Services;
 using Xunit;
 
 namespace Humans.Web.Tests.Services;
